KAMANJA-687 python server proxy integration
FacFac, Fac, and ProxyModel code complete
Python server project added with SockClient test tool, tests, doc, examples
PythonMdlSupport adjustments including compilability check for src and modelOptions parse-ability check
MetadataAPI changes...
New property in the MetadataAPI properties called PYTHON_PATH
Jpmml got changed by mistake. I update the wrong file.

To do or fix:
		s"$cmd\n$modelName\n$modelSrc"

1) The model's module name must be collected from the python file name and added to mdelem ModelDef.
This is needed for addModel.
2) The instantiation of the servers and the connections in the KamananjaMetadata area to do.
3) Filling in the node context object that is supplied to the facfac and fac with some addl values (host,port,connection object map)
4) Add facfac for python to bootstrap
5) Merge in the 1.5.0.Test version of the JpmmlFactoryOfModelInstanceFactory.scala

Status:
1) It should be possible to complete the todo items tomorrow and begin testing.
2) We need to review how we are going to integrate the python facfac into the kamanja metadata via nodecontext
3) I did not use the PROPERTIES item but added a MODELOPTIONS item in startmetadataapi.  It could be changed to
PROPERTIES if you like if in fact that is what is for. PROPERTIES seemed so generic.
4) Code review might be good
5) I am assuming that our json4s lib will allow me to dump to json string and resurrect to map or list of map.  If that
is not the case, we need to upgrade.  We will not know what the content will be like when it returns from addModel and
executeModel in particular.