KAMANJA-687 python proxy configuration
Both the metadata api and engine are configurable.

A flattened version of the following JSON string is now included and used by the PythonAdapterFactory to configure
the server(s) it will use:

PYTHON_CONFIG= {
  "PYTHON_PATH": "{InstallDirectory}/python",
  "SERVER_BASE_PORT": 8100,
  "SERVER_NODE_LIMIT": 20,
  "SERVER_HOST": "localhost",
  "PYTHON_LOG_CONFIG_PATH": "${InstallDirectory}/python/bin/pythonlog4j.cfg",
  "PYTHON_LOG_PATH": "${InstallDirectory}/python/logs/pythonserver.log"
}

If the PYTHON_CONFIG is missing from the configuration api config files supplied to our mdapi and engine,
then it is expected that the ROOT_DIR describing installation location is available and the parameters
are built from it (the above example contains the defaults).

The PYTHON_PATH contains the location of the python server .py file and its packages for the model hiearchy, commands
and the "models" package where the added models are located.

The SERVER_BASE_PORT is the first port in a band of listen ports that are used to initiate python servers on the
Kamanja node.

The SERVER_NODE_LIMIT is the maximum number of servers anticipated to be allocated.  This value should match or
exceed the number of execution threads (partition ids) to be dedicated to model execution.

The SERVER_HOST is the host where the python servers will run. By default, this is the localhost.  This is
currently thought to be the most desirable value from a Kamanja performance point of view.  However, it is
anticipated that some customer configurations might worry about more than just Kamanja and want to locate
its servers on another box where python applications/networks have been deployed.

The PYTHON_LOG_CONFIG_PATH is the name of the log4j configuration used by the python servers that are
instantiated.  With the current mechanism, all servers will write to one log file that has rollover.
Conceivably, we may be able to use the same log4j configuration as Kamanja does.  What is unknown at
this point is whether the python 2.7.x log4j implementation has the same feature set as the scala
version we use on the Kamanja platform.

The PYTHON_LOG_PATH is the name of the log file.

Status:

1) testing is underway on the metadata ingestion.
2) to do - the trunk/FactoriesOfModelInstanceFactory/PythonServer/src/main/python contains the base
information needed to be installed as part of the Kamanja installation.  The various scripts that
copy things to the target install dir need to copy this folder recursively to the Kamanja_Home.
3) test the engine with the add,subtract,multiply and divide "models"
4) a more significant model needs to be crafted and added as an example.
5) There is a good bit of documentation that has been written, but it needs to be synthesized into
a few cogent sections for the online doc.
6) Additional test model scenarios should be developed and added to the testing part of
the trunk/FactoriesOfModelInstanceFactory/PythonServer or
trunk/FactoriesOfModelInstanceFactory/PythonFactoryOfModelInstanceFactory or both.