KAMANJA-687 ModelFile => Module and name fixes
The module name and model name were not being handled properly... producing a failed
load on the pythonserver.  This is fixed.  The module and model name are now preserved
intact for use by the pythonserver by putting them in the physicalName.  That is not
folded, spindled or mutilated by the metadata api ingestion.

Instead of passing the python file name in the arguments to the python server on the
addModel command, the *moduleName* is now passed.  The server's addModel command
uses it to form the file name

Status:

With this change models are adding to the python server.  There remains an issue
on executeModel path. The new ModelInstance execute method is failing with:

ERROR [pool-17-thread-1] - Failed to execute model:add.addtuple
com.ligadata.Exceptions.NotImplementedFunctionException: execute method is not implemented for model:add.addtuple
InputMessages (1) are:org.kamanja.arithmetic.arithmeticMsg and triggerdSetIndex:0
        at com.ligadata.KamanjaBase.ModelInstance.execute(ModelBase.scala:619) ~[KamanjaInternalDeps_2.11-1.5.0.jar:0.1-SNAPSHOT]

This suggests that the engine is not liking the model def or something... looking into it
Once fixed, we should producing output (or perhaps after I fix the output handling :-).

Note that one of the differences perhaps is I am supplying an output message to the model definitions